requirements.txt
aiogram==3.17.0
aiosqlite==0.20.0
APScheduler==3.11.0
matplotlib==3.9.0
numpy==2.0.0
python-dotenv==1.0.1
Pillow==10.4.0
pytz==2024.1
